requirement.txt
streamlit
requests
python-dotenv
openai
boto3